test: consolidate 19 hand-rolled recursive rmdir helpers into RemoveDirTrait
