+++ tests/php/Integration/Run/Command/AgentInstall/AgentInstallCommandTest.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace PhelTest\Integration\Run\Command\AgentInstall;
 
 use Phel\Run\Infrastructure\Command\AgentInstallCommand;
+use PhelTest\Support\RemoveDirTrait;
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\ArrayInput;
@@ -12,6 +13,8 @@ use Symfony\Component\Console\Output\BufferedOutput;
 
 final class AgentInstallCommandTest extends TestCase
 {
+    use RemoveDirTrait;
+
     private string $testDir;
 
     private string $originalCwd;
@@ -27,7 +30,7 @@ final class AgentInstallCommandTest extends TestCase
     protected function tearDown(): void
     {
         chdir($this->originalCwd);
-        $this->removeDirectory($this->testDir);
+        $this->removeDir($this->testDir);
     }
 
     public function test_install_all_creates_every_skill_target(): void
@@ -307,34 +310,4 @@ final class AgentInstallCommandTest extends TestCase
         return new AgentInstallCommand()->run(new ArrayInput($args), $output ?? new BufferedOutput());
     }
 
-    private function removeDirectory(string $dir): void
-    {
-        if (!is_dir($dir)) {
-            return;
-        }
-
-        $items = scandir($dir);
-        if ($items === false) {
-            return;
-        }
-
-        foreach ($items as $item) {
-            if ($item === '.') {
-                continue;
-            }
-
-            if ($item === '..') {
-                continue;
-            }
-
-            $path = $dir . '/' . $item;
-            if (is_dir($path) && !is_link($path)) {
-                $this->removeDirectory($path);
-            } else {
-                unlink($path);
-            }
-        }
-
-        rmdir($dir);
-    }
 }
